refactor Directory
