+++ src/stores/directory.tsx
@@ -3,8 +3,8 @@ import urljoin from 'url-join';
 
 import { restClient } from '~/utils/rest-client';
 import { PaginationResult } from '~/interfaces/paginationResult';
-import { Directory } from '~/interfaces/directory';
-import { DirectoryTree } from '~/interfaces/directoryTree';
+import { Directory } from '~/domains/Directory';
+import { DirectoryTree } from '~/domains/DirectoryTree';
 import { useAuthenticationSWR } from '~/stores/use-authentication-swr';
 
 export const useDirectoryListSWR = (limit = 30): SWRResponse<PaginationResult<Directory>, Error> => {
